use temp file for public key for module-lwe
it appears passing the entire public key (~390kb) as a command line argument to the `module-lwe` binary is just too big.

even if it worked, it's probably not a good idea.

instead, the binary can now optionally take a file. create a temporary file with the public key, use that, and delete it when finished.

for ring-lwe, it works, so just leave it as is.